@@ -9,9 +9,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Increment_Length_Of_The_List_When_Element_Added()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(2);
+            var list = new DoublyLinkedList<int> { 1, 2 };
 
             var actualLength = list.Length;
 
@@ -21,9 +19,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Return_Element_At_The_Specified_Position()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(5);
-            list.Add(6);
+            var list = new DoublyLinkedList<int> { 5, 6 };
 
             var actualElement = list.ElementAt(1);
 
@@ -43,9 +39,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Throw_IndexOutOfRangeException_If_Index_Bigger_Or_Equal_Than_Length()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(3);
+            var list = new DoublyLinkedList<int> { 1, 3 };
 
             Action action = () => list.ElementAt(2);
 
@@ -55,9 +49,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Throw_IndexOutOfRangeException_If_Index_Less_Than_Zero()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(3);
+            var list = new DoublyLinkedList<int> { 1, 3 };
 
             Action action = () => list.ElementAt(-2);
 
@@ -67,10 +59,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Insert_Element_At_Specified_Position()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(3);
-
+            var list = new DoublyLinkedList<int> { 1, 3, 7, 9 };
             list.AddAt(1, 5);
             var actualElement = list.ElementAt(1);
 
@@ -80,8 +69,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Increment_Length()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
+            var list = new DoublyLinkedList<int> { 1 };
 
             list.AddAt(1, 5);
             var actualElement = list.Length;
@@ -92,23 +80,16 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Insert_Element_At_The_End_Of_The_List()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(6);
-
+            var list = new DoublyLinkedList<int> { 1, 6 };
             list.AddAt(2, 8);
             var actualElement = list.ElementAt(2);
-
             actualElement.Should().Be(8);
         }
 
         [Fact]
         public void Should_Insert_Element_At_The_Beginning_Of_The_List()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(6);
-
+            var list = new DoublyLinkedList<int> { 1, 6 };
             list.AddAt(0, 8);
             var actualElement = list.ElementAt(0);
 
@@ -118,9 +99,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Throw_IndexOutOfRangeException_If_Index_Bigger_Than_Length()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(6);
+            var list = new DoublyLinkedList<int> { 1, 6 };
 
             Action action = () => list.ElementAt(2);
 
@@ -128,24 +107,19 @@ namespace Tasks.Tests
         }
 
         [Fact]
-        public void Should_Remove_First_Occurance_In_The_List()
+        public void Should_Remove_First_Occurrence_In_The_List()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(2);
-            list.Add(1);
+            var list = new DoublyLinkedList<int> { 1, 2, 1 };
 
             list.Remove(1);
-
-            list.Should().BeEquivalentTo(new[] { 2, 1 });
+            list.Should()
+            .BeEquivalentTo(new[] { 2, 1 });
         }
 
         [Fact]
         public void Should_Decrement_Length_Of_The_List_If_Element_Removed()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(2);
+            var list = new DoublyLinkedList<int> { 1, 2 };
 
             list.Remove(2);
             var actualLength = list.Length;
@@ -156,9 +130,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Not_Change_Collection_If_It_Does_Not_Contain_Specified_Item()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(2);
+            var list = new DoublyLinkedList<int> { 1, 2 };
 
             list.Remove(4);
 
@@ -168,9 +140,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Not_Decrement_Length_Of_The_List_If_It_Does_Not_Contain_Specified_Item()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(2);
+            var list = new DoublyLinkedList<int> { 1, 2 };
 
             list.Remove(4);
             var actualLength = list.Length;
@@ -181,9 +151,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Remove_Element_At_The_Specified_Position()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(5);
-            list.Add(6);
+            var list = new DoublyLinkedList<int> { 5, 6 };
 
             list.RemoveAt(1);
 
@@ -193,9 +161,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Return_Removing_Item()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(2);
-            list.Add(5);
+            var list = new DoublyLinkedList<int> { 2, 5 };
 
             var removedItem = list.RemoveAt(1);
 
@@ -205,9 +171,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Decrement_Length_Of_The_List_If_Element_Removed_Using_RemoveAt()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(1);
-            list.Add(2);
+            var list = new DoublyLinkedList<int> { 1, 6 };
 
             list.RemoveAt(1);
             var actualLength = list.Length;
@@ -228,9 +192,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Throw_IndexOutOfRangeException_If_Index_For_RemoveAt_Bigger_Or_Equal_Than_Length()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(2);
-            list.Add(5);
+            var list = new DoublyLinkedList<int> { 2, 5 };
 
             Action action = () => list.RemoveAt(2);
 
@@ -240,9 +202,7 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Throw_IndexOutOfRangeException_If_Index_For_RemoveAt_Less_Than_Zero()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(2);
-            list.Add(5);
+            var list = new DoublyLinkedList<int> { 2, 5 };
 
             Action action = () => list.RemoveAt(-1);
 
@@ -252,12 +212,9 @@ namespace Tasks.Tests
         [Fact]
         public void Should_Iterate_In_ForEach_Loop()
         {
-            var list = new DoublyLinkedList<int>();
-            list.Add(2);
-            list.Add(5);
-            list.Add(42);
+            var list = new DoublyLinkedList<int> { 2, 5, 42 };
 
-            int i = 0;
+            var i = 0;
             foreach (var item in list)
             {
                 item.Should().Be(list.ElementAt(i));
